% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CentileGraph.R
\name{calculateCentile}
\alias{calculateCentile}
\title{calculateCentile
This function calculates centiles from the days of cycle and values stores in object.}
\usage{
calculateCentile(obj)
}
\arguments{
\item{obj}{CentileGraph object}
}
\value{
CentileGraph object with centiles.
}
\description{
calculateCentile
This function calculates centiles from the days of cycle and values stores in object.
}
