% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CentileGraph.R
\name{calculateQuantiles}
\alias{calculateQuantiles}
\title{Calculation of quantiles}
\usage{
calculateQuantiles(
  obj,
  days_of_cycle = 6:12,
  quantiles = c(0, 0.25, 0.5, 0.75, 1),
  model = "beta"
)
}
\arguments{
\item{obj}{CentileGraph object}

\item{days_of_cycle}{days of cycle to calculate quantiles for.}

\item{quantiles}{Set of quantiles to calculate. Default c(0,0.25,0.50,0.75,1)}

\item{model}{Model distribution. Default "beta".}
}
\value{
Dataframe with quantiles per cycle day.
}
\description{
Calculation of quantiles
}
\examples{
calculateQuantiles
}
