% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CentileGraph.R
\name{createCentileGraph}
\alias{createCentileGraph}
\title{Create a CentileGraph object}
\usage{
createCentileGraph(cycle_day, value)
}
\arguments{
\item{cycle_day}{Vector with days of cycle (LH+)}

\item{value}{Vector of raw values to centilise.}
}
\value{
A CentileGraph object
}
\description{
This function initialises a CentileGraph object.
It requires days of cycle and value of assay (uNK score, RT-qPCR expression).
}
