% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CentileGraph.R
\name{plotCentiles}
\alias{plotCentiles}
\title{Title}
\usage{
plotCentiles(
  obj,
  file = NULL,
  days_of_cycle = 6:12,
  assay,
  width = 6,
  height = 6
)
}
\arguments{
\item{obj}{CentileGraph object after quantile calculation}

\item{file}{Path to location where pdf will be saved. If blank, ggplot2 object will be returned.}

\item{days_of_cycle}{Vector of days of cycle to plot.}

\item{assay}{Name of value column.}

\item{width}{Width of pdf if saving to file.}

\item{height}{Height of pdf if saving to file.}
}
\value{
ggplot2 object or PDF file generated.
}
\description{
Title
}
\examples{
plotCentiles(qupath,"test.pdf")
}
